aleartservice/requirements.txt
asgiref==3.4.1
blinker==1.4
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
ebaysdk==2.2.0
gunicorn==20.1.0
idna==3.2
install==1.3.4
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.9.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
whitenoise==5.3.0